fastapi-desafio/src/requirements.txt
psycopg2-binary==2.9.5
SQLAlchemy==1.4.41
fastapi==0.87.0
uvicorn==0.20.0
requests==2.31.0
bcrypt==4.2.0
python-jose==3.3.0
passlib==1.7.4
alembic==1.7.7

# dev
pytest==7.2.0
httpx==0.23.1